test(gpu): consolidate CUDA runtime adapter tests
